Refactored the Code a little bit
Added new classes for noteManager, utility class, and exported
the Note class in another file.

The utility class provides a simple method for creating new elements
and adding classes and content to them

The Note class now uses the utility class to create the note elements
the noteManager is not being used yet but will be used in the future.